Rename : data -> domain.entity로 이동
data.DiaryDTO -> domain.entity.diary로 이동
data.Emotion, EmotionDTO -> domain.entity.emotiond으로 이동
나머지 -> 이동에 따른 변경
