+++ app/src/main/res/layout/item_diary.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="diaryContent"
-            type="konkuk.gdsc.memotion.data.DiarySimple" />
+            type="konkuk.gdsc.memotion.domain.entity.diary.DiarySimple" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
